Add initial Svelte parser implementation
Introduces a new svelte-parser crate with core parsing modules for Svelte fragments, elements, attributes, comments, and expressions. Updates dependencies to include swc_ecma_parser and related crates for JS/TS parsing. Refactors svelte-ast span fields to use usize for better compatibility. Adds comprehensive parsing logic and tests for Svelte syntax.

+++ crates/svelte-ast/src/span.rs
@@ -3,12 +3,12 @@ use serde::Serialize;
 /// Source position span (byte offsets into the original .svelte source).
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize)]
 pub struct Span {
-    pub start: u32,
-    pub end: u32,
+    pub start: usize,
+    pub end: usize,
 }
 
 impl Span {
-    pub fn new(start: u32, end: u32) -> Self {
+    pub fn new(start: usize, end: usize) -> Self {
         Self { start, end }
     }
